Places the generateSequence function and the _maxNumberOfCharactersToPrint private field from the sequences module to a new module called generator
